knowledge-base/requirements.txt
# 知识库 - Python 依赖

# 向量数据库
qdrant-client==1.7.0

# 嵌入模型
sentence-transformers==2.2.2
# sentence-transformers 2.2.2 仍依赖 cached_download，需避免 huggingface_hub 新版不兼容
huggingface-hub<0.26
# 中文优化模型
FlagEmbedding==1.2.5
torch>=2.0.0

# 文档处理
langchain==0.1.0
langchain-community==0.0.10
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1

# 代码解析
tree-sitter==0.21.3
tree-sitter-python>=0.23,<0.24
tree-sitter-javascript>=0.23,<0.24

# Web 框架
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# 工具
numpy==1.26.2
scikit-learn==1.3.2
redis==5.0.1

# 日志
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.23.2

# 开发
python-dotenv==1.0.0
